% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseaseEnrichment-package.R
\docType{package}
\name{diseaseEnrichment-package}
\alias{diseaseEnrichment}
\alias{diseaseEnrichment-package}
\title{diseaseEnrichment: Disease Enrichment}
\description{
Disease enrichment from a list of genes.
}
\author{
\strong{Maintainer}: Naomi Yudanin \email{nyudanin@envisagenics.com}

}
\keyword{internal}
